feat(api): return auth tokens in auth responses
We cannot send multiple cookies in the response, so we can't really save
the tokens in the cookies. With cookies, it would be rather easy to set
up auth, but now we need to figure out something else. For now, we just
return the tokens in the response body and have the client pass them in
as Authorization headers.

Some references:

- https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie
- https://www.github.com/whatwg/fetch/issues/506
- https://www.github.com/apollographql/router/issues/2211

Related to #28