feat(api): add mutation for logging out
This should not be required in normal cases as the client JS could just
delete the auth cookies. However, since the cookies are HTTPOnly, then
they cannot be deleted by client-side JS and need to be deleted on the
server.

I also considered adding a third cookie that not HTTPOnly that could be
deleted on the client side without making a mutation, but decided
against that for now. It should be a valid workaround that would also
work in offline mode.

Related to #28