fix(api): do not fail when no seasons to sync
In some cases (e.g., https://www.imdb.com/title/tt7708606/), the total
seasons number from OMDb is "N/A", which breaks the `Array.from`
behavior as `parseInt('N/A')` results in `NaN`. So, we just ignore those
cases for now. Maybe in the future we'll add a way to add episodes and
seasons to such series manually.

Related to #30