fix(webapp): move current user query to server file
The problem was that the current user was not available in route
`.server.ts` `onBeforeRender` functions. Moving the current user to the
initial context that's used to render the HTML seems to fix the issue.
The `_default.page.server.tsx` code is executed AFTER `onBeforeRender`
functions of routes.

Related to #43