fix(api): do not fail when trying to insert the same season twice
Basically, the series details request is sometimes triggered twice, once
on the server side, and a second time on the client side. It can happen
that these are very close together and there might be a race condition
so that both try to sync the series details. The first one succeeds, but
the second one tries to insert already existing seasons into the
database.