fix(webapp): disable form elements until hydration
The elements should be disabled on the server and re-enabled once the
page is hydrated. This is, for example, because clicking on a submit
button before the JS has been loaded, results in a native submit event.
We don't want to allow the user to interact with the page until it is
actually interactive.

This is an even bigger problem in UI tests, where the tests are quite
fast and will probably click on buttons before hydration is done.
Playwright is smart enough to wait for buttons to become interactive
before clicking on them.

The `autoComplete="off"` hack is because of Firefox, which saves and
restores the form element values (and disabled states!) when the page is
refreshed. This results in hydration errors because the restored HTML
does not have the button as disabled, but the hydrated HTML DOES
(because `isSSR` is initially `true` for 1 render). This is further
discussed here: https://github.com/vercel/next.js/discussions/21999

More resources:
- https://stackoverflow.com/questions/5985839/bug-with-firefox-disabled-attribute-of-input-not-resetting-when-refreshing
- https://bugzilla.mozilla.org/show_bug.cgi?id=654072

Related to #44