fix(webapp): fix favicon (again)
Hopefully this time it will work...

The `favicon.ico` file did not get outputted into the `dist` folder when
running `build`. Probably because it was not actually imported anywhere
in the CLIENT code (assets are only available from `dist/client`).
Adding a dummy import into the client default page seems to work, but
it's probably not a good solution.

Maybe once the server code gets transpiled by Vite in vite-plugin-ssr
v1, this will be fixed?