build: fix environment variables with nx
This caused an issue that `pnpm test:coverage` did not target the
correct database.

`nx` loads environment variables from the `.env` files based on the task
name (e.g., `.env.test:coverage`) automatically. This means that
`dotenv-cli` call inside the app scripts themselves does not work
because the environment variables are ALREADY loaded and won't be loaded
again. Basically, `.env` is loaded by `nx` and then `dotenv-cli` is
triggered, which reads the `.env.test` file, but does not set any
variables because they are already set.

Related to #60