feat(api): delete series if it has been deleted from TMDB
This requires some refactoring of the TMDB request making as we want to
differentiate between a not found response and if parsing fails. If the
parsing fails, then that is not enough cause to delete the series as
there might just be some random breaking change that we would need to
handle. In the future, we should probably send a Sentry error or
something when parsing fails.

Related to #72