fix(webapp): use HTTPS for series posters
This should avoid a console warning in some browsers. I think that on
HTTPS, Chromium automatically upgrades HTTP asset requests to HTTPS
anyway, so this already kinda worked with HTTPS before. Now it's just
"official".

The HTTPS poster URLs also seem to be working locally with HTTP, so
that's great.