fix(webapp): remove delay from text color transitions
This is very similar to the previous commit fixing the icon colors.
Basically, we're setting the text color on EACH element rather than only
on the body and letting it cascade down. This means that the CSS is less
efficient, BUT it avoids delays when transitioning the colors of nested
text elements.

Helpful links:

https://codepen.io/zeizig/pen/gOZqPER
https://stackoverflow.com/questions/39600138/how-to-simultaneously-transition-colour-of-nested-children