build: use esbuild to bundle server code
A problem was that in production mode (compiled to ESM JS and running
with `node`), importing internal packages (`@serieslist/*`) pointed to
the TypeScript files. This meant that importing those packages DID NOT
WORK (e.g., `@serieslist/logger`). `tsc` does not bundle any
dependencies.

To fix this, we can bundle the code while building, bundling the
`@serieslist/*` packages INTO the `dist/` folder and marking all other
direct dependencies as external. This means that we can keep
`package.json` `main` and `exports` fields referencing the TypeScript
files.

I'm not super happy with the bundling, though, as there are some weird
things going on with ESM. I considered these options for the bundler:
- `esbuild`
- Vite with `vite-node` might be nice as we're already using Vitest
- `swc`
- `webpack`

The packages themselves are never built on their own. In order to build
a package, one of the consuming apps needs to be built and pull the
package code into the apps dist folder.

This seems to be the approach that's mostly recommended for smaller and
medium sized projects.

Some more resources:
- https://turbo.build/repo/docs/handbook/sharing-code/internal-packages
- https://turbo.build/blog/you-might-not-need-typescript-project-references
- https://www.reddit.com/r/typescript/comments/10ebgbs/handling_typescript_in_a_monorepo/
- https://nx.dev/concepts/more-concepts/incremental-builds

I also tried to build the packages themselves and point `main` and
`exports` to the compiled JS files in `dist/`, but that caused a couple
issues like Go To Definition going to the type definition in `dist/`,
not to the source code (which is super inconvenient). It would also mean
that each package needs its own build step with configuration, which
would be a bit more painful to set up.

Related to #83