build: fix filenames of migrations in production bundle
Previously, they were generated with the same name as the source (with
colons), but now, when building with `esbuild`, colons are changed to
underscores. This change renames the files back to being with colons.
The migrations created in the future should use dashes instead of colons
to avoid weird workarounds being required in the future.

Related to #83