build(tmdb): mark as not having side-effects
It turns out that if API is built and one of the internal monorepo
dependencies imports some files that include `devDependencies`, then an
import of that dependency is still included in the built bundle. This
happens even if there are no `sideEffects` in the importing application.
We ALSO need to set `sideEffects: false` for the package.

Related to #83