build(api): do not use async import as it bundles everything
Using an async import causes all of the imported things to be added to
the bundle, ignoring the `sideEffects` settings in `package.json` files
of the packages. This meant that we were including `nock` and `fishery`
from `packages/tmdb` in the bundle...

Since everything is bundled into one file anyways, then the dynamic
import doesn't actually do anything useful.

Related to #83