fix(e2e): change order of series progress and status change
Seems like there might be a race condition happening. In some e2e test
runs, the response for toggling an episode as seen is EMPTY.

This doesn't make much sense, but I'm guessing that we're moving away
from the page too fast after marking an episode as seen.

This fix seems to work locally, let's see if it also works in CI.